JPAR-116 - define dummy but real versions for Shibboleth components
This means that the Javadoc links included in the configuration will always resolve to something real, even in projects like java-support which do not use the dependencies (and thus have their own definition of the version properties).

Projects MUST NOT rely on these being "real" or "correct" if they do have a dependency, however: in that case they MUST define the properties themselves.