IDP-1465 - Make dependency on surefire-testng explicit
The maven-surefire-plugin has dynamic dependencies on providers dependent on the test frameworks detected in the test classpath. If the provider isn't available for download (e.g., in the central-disabled environment nightly Jenkins jobs run in) then you just get a class not found exception.

Work round this by making the dependency explicit for the provider we need.

https://issues.shibboleth.net/jira/browse/IDP-1465